requirements.txt
Django>=5.0
django-tenants
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
gunicorn
reportlab
pdfplumber